- Fixed split UTXO function names & params
