+++ lib/Wallet.js
@@ -281,157 +281,141 @@ class Wallet {
     verifyMessage(msg, address, signature) {
         return bitcoinjs_message_1.default.verify(msg, address, signature, "\u0016HYDRA Signed Message:\n");
     }
-    // public async optimizeUTXOS(
-    //   utxos: IUTXO[],
-    //   keyPair: ECPair,
-    //   feeRate: number,
-    //   utxoMinValue: number,
-    //   utxoThreshold: number
-    // ): Promise<{ hex: string; error: string }> {
-    //   var tx = new TransactionBuilder(keyPair.network);
-    //   var fee: BigNumber;
-    //   const from: string = this.address;
-    //   let walletBalance: BigNumber = sumUTXOs(utxos);
-    //   if (walletBalance.lte(new BigNumber(utxoMinValue).times(1e8))) {
-    //     return {
-    //       hex: "",
-    //       error: `User's Balance is below or equal to the threshold of ${utxoMinValue} HYDRA. No UTXOs to optimize.`,
-    //     };
-    //   }
-    //   var validUTXOs = filterUtxos(utxos, utxoThreshold);
-    //   if (validUTXOs.length == 0) {
-    //     return {
-    //       hex: "",
-    //       error: "No UTXOs to optimize.",
-    //     };
-    //   }
-    //   var balanceOfValidUTXOS: BigNumber = sumUTXOs(validUTXOs);
-    //   // get all  of the UTXOSs above the threshold  of the wallet
-    //   var UTXOSofWalletAboveThreshold: IUTXO[] = utxos.filter((utxo: any) => {
-    //     const value = new BigNumber(utxo.value);
-    //     if (
-    //       value.gt(new BigNumber(utxoThreshold).times(1e8)) /// upper threshold for utxo filter is now 150
-    //     ) {
-    //       return true;
-    //     }
-    //     return false;
-    //   });
-    //   if (balanceOfValidUTXOS.gt(new BigNumber(utxoThreshold).times(1e8))) {
-    //     fee = calculateFee(validUTXOs, 1, feeRate, keyPair);
-    //     addInputsOutputsAndSignInputs(validUTXOs, balanceOfValidUTXOS);
-    //     return {
-    //       hex: tx.build().toHex(),
-    //       error: "",
-    //     };
-    //   } else if (
-    //     balanceOfValidUTXOS.gt(new BigNumber(utxoMinValue).times(1e8)) &&
-    //     UTXOSofWalletAboveThreshold.length === 0
-    //   ) {
-    //     fee = calculateFee(validUTXOs, 1, feeRate, keyPair);
-    //     if (fee.gt(balanceOfValidUTXOS)) {
-    //       return {
-    //         hex: "",
-    //         error: "Not enough balance to pay fee.",
-    //       };
-    //     }
-    //     addInputsOutputsAndSignInputs(validUTXOs, balanceOfValidUTXOS);
-    //     return {
-    //       hex: tx.build().toHex(),
-    //       error: "",
-    //     };
-    //   } else {
-    //     let arrayOfValidUTXOSPlusFirstUTXOAboveThreshold: IUTXO[] = [
-    //       ...validUTXOs,
-    //       UTXOSofWalletAboveThreshold[0],
-    //     ];
-    //     fee = calculateFee(
-    //       arrayOfValidUTXOSPlusFirstUTXOAboveThreshold,
-    //       1,
-    //       feeRate,
-    //       keyPair
-    //     );
-    //     if (fee.gt(balanceOfValidUTXOS)) {
-    //       return {
-    //         hex: "",
-    //         error: "Not enough balance to pay fee.",
-    //       };
-    //     }
-    //     var balanceOfValidUTXOSPlusFirstUTXOAboveThreshold = sumUTXOs(
-    //       arrayOfValidUTXOSPlusFirstUTXOAboveThreshold
-    //     );
-    //     addInputsOutputsAndSignInputs(
-    //       arrayOfValidUTXOSPlusFirstUTXOAboveThreshold,
-    //       balanceOfValidUTXOSPlusFirstUTXOAboveThreshold
-    //     );
-    //     return {
-    //       hex: tx.build().toHex(),
-    //       error: "",
-    //     };
-    //   }
-    //   function calculateFee(
-    //     inputs: IUTXO[],
-    //     outputs: any,
-    //     feeRate: any,
-    //     keyPair: ECPair
-    //   ) {
-    //     var tx = new TransactionBuilder(keyPair.getNetwork());
-    //     for (var i = 0; i < inputs.length; i++) {
-    //       ///adds the inputs to the tx
-    //       tx.addInput(inputs[i].hash, inputs[i].outputIndex);
-    //     }
-    //     for (var i = 0; i <= outputs; i++) {
-    //       tx.addOutput(from, new BigNumber(utxoMinValue).times(1e8).toNumber());
-    //     }
-    //     // Sign the inputs
-    //     for (var i = 0; i < inputs.length; i++) {
-    //       tx.sign(i, keyPair);
-    //     }
-    //     return new BigNumber(tx.build().toHex().length).times(feeRate);
-    //   }
-    //   function filterUtxos(utxos: Array<IUTXO>, utxoThreshold: number) {
-    //     return utxos.filter((utxo: any) => {
-    //       const value = new BigNumber(utxo.value);
-    //       if (
-    //         value.gt(new BigNumber(25).times(1e6)) &&
-    //         value.lt(new BigNumber(utxoThreshold).times(1e8))
-    //       ) {
-    //         return true;
-    //       }
-    //       return false;
-    //     });
-    //   }
-    //   function addInputsOutputsAndSignInputs(
-    //     arrayOfUTXOSForInput: Array<IUTXO>,
-    //     ouputAmount: any
-    //   ) {
-    //     // Add the inputs
-    //     for (var i = 0; i < arrayOfUTXOSForInput.length; i++) {
-    //       tx.addInput(
-    //         arrayOfUTXOSForInput[i].hash,
-    //         arrayOfUTXOSForInput[i].outputIndex
-    //       );
-    //     }
-    //     // Add the outputs
-    //     tx.addOutput(from, ouputAmount.minus(fee).toNumber());
-    //     // Sign the inputs
-    //     for (var i = 0; i < arrayOfUTXOSForInput.length; i++) {
-    //       tx.sign(i, keyPair);
-    //     }
-    //   }
-    //   function sumUTXOs(utxos: Array<IUTXO>) {
-    //     let sum = new BigNumber(0);
-    //     for (let utxo of utxos) {
-    //       sum = sum.plus(utxo.value);
-    //     }
-    //     return sum;
-    //   }
-    // }
+    optimizeUTXOS(utxos, keyPair, feeRate, utxoMinValue, utxoThreshold) {
+        return __awaiter(this, void 0, void 0, function* () {
+            var tx = new bitcoinjs_lib_1.TransactionBuilder(keyPair.network);
+            var fee;
+            const from = this.address;
+            let walletBalance = sumUTXOs(utxos);
+            if (walletBalance.lte(new bignumber_js_1.BigNumber(utxoMinValue).times(1e8))) {
+                return {
+                    hex: "",
+                    error: `User's Balance is below or equal to the threshold of ${utxoMinValue} HYDRA. No UTXOs to optimize.`,
+                };
+            }
+            var validUTXOs = filterUtxos(utxos, utxoThreshold);
+            if (validUTXOs.length == 0) {
+                return {
+                    hex: "",
+                    error: "No UTXOs to optimize.",
+                };
+            }
+            var balanceOfValidUTXOS = sumUTXOs(validUTXOs);
+            ///////
+            // get all  of the UTXOSs above the threshold  of the wallet
+            var UTXOSofWalletAboveThreshold = utxos.filter((utxo) => {
+                const value = new bignumber_js_1.BigNumber(utxo.value);
+                if (value.gt(new bignumber_js_1.BigNumber(utxoThreshold).times(1e8)) /// upper threshold for utxo filter is now 150
+                ) {
+                    return true;
+                }
+                return false;
+            });
+            if (balanceOfValidUTXOS.gt(new bignumber_js_1.BigNumber(utxoThreshold).times(1e8))) {
+                fee = calculateFee(validUTXOs, 1, feeRate, keyPair);
+                addInputsOutputsAndSignInputs(validUTXOs, balanceOfValidUTXOS);
+                return {
+                    hex: tx.build().toHex(),
+                    error: "",
+                };
+            }
+            else if (balanceOfValidUTXOS.gt(new bignumber_js_1.BigNumber(utxoMinValue).times(1e8)) &&
+                UTXOSofWalletAboveThreshold.length === 0) {
+                fee = calculateFee(validUTXOs, 1, feeRate, keyPair);
+                if (fee.gt(balanceOfValidUTXOS)) {
+                    return {
+                        hex: "",
+                        error: "Not enough balance to pay fee.",
+                    };
+                }
+                addInputsOutputsAndSignInputs(validUTXOs, balanceOfValidUTXOS);
+                return {
+                    hex: tx.build().toHex(),
+                    error: "",
+                };
+            }
+            else {
+                let arrayOfValidUTXOSPlusFirstUTXOAboveThreshold = [
+                    ...validUTXOs,
+                    UTXOSofWalletAboveThreshold[0],
+                ];
+                fee = calculateFee(arrayOfValidUTXOSPlusFirstUTXOAboveThreshold, 1, feeRate, keyPair);
+                if (fee.gt(balanceOfValidUTXOS)) {
+                    return {
+                        hex: "",
+                        error: "Not enough balance to pay fee.",
+                    };
+                }
+                var balanceOfValidUTXOSPlusFirstUTXOAboveThreshold = sumUTXOs(arrayOfValidUTXOSPlusFirstUTXOAboveThreshold);
+                addInputsOutputsAndSignInputs(arrayOfValidUTXOSPlusFirstUTXOAboveThreshold, balanceOfValidUTXOSPlusFirstUTXOAboveThreshold);
+                return {
+                    hex: tx.build().toHex(),
+                    error: "",
+                };
+            }
+            function calculateFee(inputs, outputs, feeRate, keyPair) {
+                var tx = new bitcoinjs_lib_1.TransactionBuilder(keyPair.getNetwork());
+                for (var i = 0; i < inputs.length; i++) {
+                    ///adds the inputs to the tx
+                    tx.addInput(inputs[i].hash, inputs[i].outputIndex);
+                }
+                for (var i = 0; i <= outputs; i++) {
+                    tx.addOutput(from, new bignumber_js_1.BigNumber(utxoMinValue).times(1e8).toNumber());
+                }
+                // Sign the inputs
+                for (var i = 0; i < inputs.length; i++) {
+                    tx.sign(i, keyPair);
+                }
+                return new bignumber_js_1.BigNumber(tx.build().toHex().length).times(feeRate);
+            }
+            function filterUtxos(utxos, utxoThreshold) {
+                return utxos.filter((utxo) => {
+                    const value = new bignumber_js_1.BigNumber(utxo.value);
+                    if (
+                    //  if (value.gt(new BigNumber(4).times(1e6))) {
+                    value.gt(new bignumber_js_1.BigNumber(25).times(1e6)) &&
+                        value.lt(new bignumber_js_1.BigNumber(utxoThreshold).times(1e8))) {
+                        return true;
+                    }
+                    return false;
+                });
+            }
+            function addInputsOutputsAndSignInputs(arrayOfUTXOSForInput, ouputAmount) {
+                // Add the inputs
+                for (var i = 0; i < arrayOfUTXOSForInput.length; i++) {
+                    tx.addInput(arrayOfUTXOSForInput[i].hash, arrayOfUTXOSForInput[i].outputIndex);
+                }
+                // Add the outputs
+                tx.addOutput(from, ouputAmount.minus(fee).toNumber());
+                // Sign the inputs
+                for (var i = 0; i < arrayOfUTXOSForInput.length; i++) {
+                    tx.sign(i, keyPair);
+                }
+            }
+            function sumUTXOs(utxos) {
+                let sum = new bignumber_js_1.BigNumber(0);
+                for (let utxo of utxos) {
+                    sum = sum.plus(utxo.value);
+                }
+                return sum;
+            }
+        });
+    }
+    optimizeWalletUTXOS(utxoMinValue, utxoThreshold) {
+        return __awaiter(this, void 0, void 0, function* () {
+            const utxos = yield this.getBitcoinjsUTXOs();
+            const infoRes = yield this.insight.getBlockchainInfo();
+            const feeRate = Math.ceil(infoRes.feeRate * 1e5);
+            let txResponse = yield this.optimizeUTXOS(utxos, this.keyPair, feeRate, utxoMinValue || 100, utxoThreshold || 150);
+            return txResponse.hex !== ""
+                ? yield this.sendRawTx(txResponse.hex)
+                : txResponse.error;
+        });
+    }
     splitUTXOS(utxos, keyPair, feeRate, utxoMinValue, utxoMaxValue, resultingUTXO //resultingUTXO
     ) {
         return __awaiter(this, void 0, void 0, function* () {
-            // Filter the utxos so that they don't have value less than 0.04 HYDRA or 100
             var validUTXOs = filterUtxosOutsideRange(utxos, utxoMinValue, utxoMaxValue);
-            console.log("[4):SPLIT UTXO::LOG:]:- Valid UTXOs:", validUTXOs);
             if (validUTXOs.length == 0) {
                 return {
                     hex: "",
@@ -446,7 +430,6 @@ class Wallet {
             });
             // Calculate the total balance
             var balance = sumUTXOs(validUTXOs);
-            console.log("[5):SPLIT UTXO::LOG:]:- Balance:", balance);
             if (balance.lte(new bignumber_js_1.BigNumber(resultingUTXO).times(1e8))) {
                 return {
                     hex: "",
@@ -457,27 +440,22 @@ class Wallet {
             var outputs;
             // Calculate the total outputs with value of resultingUTXO
             var totalOutputs = balance.dividedToIntegerBy(new bignumber_js_1.BigNumber(resultingUTXO).times(1e8));
-            console.log("[6):SPLIT UTXO::LOG:]:- Total Outputs:", totalOutputs.toNumber());
             if (totalOutputs.toNumber() > 100) {
                 // If more than 100 => therefore the balance is over 10000 => 100 outputs with 100 and 1 with the change
                 outputs = 101;
                 validUTXOs = selectUTXOs(validUTXOs, feeRate, new bignumber_js_1.BigNumber(resultingUTXO).times(1e8).times(100).toNumber(), outputs);
-                console.log("[7):SPLIT UTXO::LOG:]:- Valid UTXOs Length:", validUTXOs.length);
                 balance = sumUTXOs(validUTXOs);
             }
             else {
                 outputs = totalOutputs.toNumber();
             }
             var tx = new bitcoinjs_lib_1.TransactionBuilder(keyPair.network);
-            console.log("[8):SPLIT UTXO::LOG:]:- outputs:", outputs);
             // Get the walet address
             var from = this.address;
             //Calculate the value for the transaction
             var value = new bignumber_js_1.BigNumber(resultingUTXO).times(1e8).times(outputs);
             // Calculate fee with the current inputs and outputs
             fee = calculateFee(validUTXOs, outputs, feeRate, keyPair);
-            console.log("[13):SPLIT UTXO::LOG:]:- Valid UTXOs Length:", validUTXOs.length);
-            console.log("[14):SPLIT UTXO::LOG:]:- Fees:", fee.toNumber());
             if (fee.gt(balance)) {
                 return {
                     hex: "",
@@ -516,18 +494,7 @@ class Wallet {
                 hex: tx.build().toHex(),
                 error: "",
             };
-            //////////////////SUPPORT FUNCTIONS/////////////////////
-            function filterUtxos(utxos, resultingUTXO) {
-                return utxos.filter((utxo) => {
-                    const value = new bignumber_js_1.BigNumber(utxo.value);
-                    if (value.gt(new bignumber_js_1.BigNumber(4).times(1e6))) {
-                        if (!value.eq(new bignumber_js_1.BigNumber(resultingUTXO).times(1e8))) {
-                            return true;
-                        }
-                    }
-                    return false;
-                });
-            }
+            //////////////////SUPPORT FUNCTIONS  [SPLIT UTXO] /////////////////////
             function filterUtxosOutsideRange(utxos, utxoMinValue, utxoMaxValue) {
                 return utxos.filter((utxo) => {
                     const value = new bignumber_js_1.BigNumber(utxo.value);
@@ -572,20 +539,16 @@ class Wallet {
                 for (var i = 0; i < inputs.length; i++) {
                     tx.sign(i, keyPair);
                 }
-                console.log("[9):SPLIT UTXO::LOG:]:-[fn() - calculateFee]::- feeRate", feeRate);
-                console.log("[10):SPLIT UTXO::LOG:]:-[fn() - calculateFee]::- tx.build().toHex().length", tx.build().toHex().length);
                 // Calculate the total fee
                 const totalFee = new bignumber_js_1.BigNumber(tx.build().toHex().length).times(feeRate);
-                console.log("[11):SPLIT UTXO::LOG:]:- Total Fee:", totalFee.toNumber());
                 // Calculate and return half of the fee
                 const halfFee = totalFee.dividedBy(2);
-                console.log("[12):SPLIT UTXO::LOG:]:- Half Fee:", halfFee.toNumber());
                 return halfFee;
             }
-            //////////////////SUPPORT FUNCTIONS/////////////////////
+            //////////////////SUPPORT FUNCTIONS [SPLIT UTXO]/////////////////////
         });
     }
-    optimizeWalletUTXOS(utxoMinValue, utxoMaxValue, utxoThreshold // resultingUTXO
+    splitWalletUTXOS(utxoMinValue, utxoMaxValue, utxoThreshold // resultingUTXO
     ) {
         return __awaiter(this, void 0, void 0, function* () {
             try {
@@ -593,15 +556,10 @@ class Wallet {
                 const utxos = yield this.getBitcoinjsUTXOs();
                 // Retrieve blockchain information to get the fee rate
                 const infoRes = yield this.insight.getBlockchainInfo();
-                // Log the UTXOs for reference
-                console.log("[1):SPLIT UTXO::LOG:]:- UTXOs:", utxos);
                 // Calculate the fee rate for the transaction
                 const feeRate = Math.ceil(infoRes.feeRate * 1e5);
-                console.log("[2):SPLIT UTXO::LOG:]:- (Math.ceil(infoRes.feeRate * 1e5))Fee Rate:", feeRate);
                 // Split UTXOs based on specified parameters
-                let txResponse = yield this.splitUTXOS(utxos, this.keyPair, feeRate, utxoMinValue || 100, utxoMaxValue || 110, utxoThreshold || 150);
-                // Log the result of UTXO splitting
-                console.log("[(FINAL):SPLIT UTXO::LOG:]:- Split UTXOs Result:", txResponse);
+                let txResponse = yield this.splitUTXOS(utxos, this.keyPair, feeRate, utxoMinValue || 0.001, utxoMaxValue || 200, utxoThreshold || 100);
                 // If a valid transaction hex is generated, send the transaction
                 return txResponse.hex !== ""
                     ? yield this.sendRawTx(txResponse.hex)
